+++ server/src/main/java/com/kalyvianakis/estiator/io/service/ITableService.java
@@ -8,8 +8,8 @@ import java.util.List;
 public interface ITableService {
     public Table save(Table table);
     public List<Table> get();
-    public Table get(Integer id) throws ResourceNotFoundException;
-    public void delete(Integer id) throws ResourceNotFoundException;
-    public Boolean exists(Integer id);
-    public Boolean notExists(Integer id);
+    public Table get(Long id) throws ResourceNotFoundException;
+    public void delete(Long id) throws ResourceNotFoundException;
+    public Boolean exists(Long id);
+    public Boolean notExists(Long id);
 }
